Remove Error Struct > change to type, use min_cl
