+++ Cargo.toml
@@ -13,6 +13,7 @@ rust-version = "1.65"
 
 [dependencies]
 #cuwanto-server = {path="../cuwanto/cuwanto-server", optional=true}
+min-cl = { path = "../min-cl", optional=true }
 
 # WGPU deps
 futures-intrusive = {version="0.5.0", optional=true}
@@ -25,7 +26,7 @@ custos = { version="0.5.0", default-features = false, optional=true }
 
 [features]
 default = ["blas", "stack", "static-api", "opencl", "wgpu", "cuda"]
-opencl = ["custos/opencl"]
+opencl = ["custos/opencl", "dep:min-cl"]
 #network = ["cuwanto-server"]
 cuda = []
 realloc = []
