+++ build.rs
@@ -1,3 +1,4 @@
+
 fn main() {
     // TODO: execute other opencl test to know whether opencl can actually be used
     #[cfg(feature = "opencl")]
