--- src/devices/opencl/api/mod.rs
@@ -1,7 +1,0 @@
-pub use cl::*;
-pub use error::*;
-pub use ffi::*;
-
-mod cl;
-mod error;
-mod ffi;
