+++ tests/error.rs
@@ -3,7 +3,8 @@ use custos::{DeviceError, Error};
 #[cfg(feature = "opencl")]
 #[test]
 fn test_error() {
-    use custos::{opencl::api::OCLErrorKind, OpenCL};
+    use custos::{OpenCL, ErrorKind};
+    use min_cl::api::OCLErrorKind;
 
     let device = OpenCL::new(1000000000000000000);
 
@@ -33,7 +34,7 @@ fn test_error_panics() {
 
 #[cfg(feature = "opencl")]
 #[test]
-fn test_questionmark() -> Result<(), Box<dyn std::error::Error>> {
+fn test_questionmark() -> Result<(), Box<dyn std::error::Error + Sync+Send>> {
     use custos::OpenCL;
 
     let _device = OpenCL::new(0)?;
@@ -56,9 +57,8 @@ fn test_print_error() {
 #[test]
 fn test_std_err() {
     let err = Error::from(DeviceError::ConstructError);
-    let e: Box<dyn std::error::Error> = err.into();
     assert_eq!(
-        e.downcast_ref::<DeviceError>(),
+        err.downcast_ref::<DeviceError>(),
         Some(&DeviceError::ConstructError)
     );
 }
@@ -66,7 +66,8 @@ fn test_std_err() {
 #[cfg(feature = "opencl")]
 #[test]
 fn test_ocl_errors() {
-    use custos::opencl::api::OCLErrorKind;
+    use min_cl::api::OCLErrorKind;
+
 
     for i in -70..=-1 {
         let err = OCLErrorKind::from_value(i);
